requirements.txt
altair==4.2.0
attrs==22.1.0
backports.zoneinfo==0.2.1
blinker==1.5
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
contourpy==1.0.5
cryptography==38.0.1
cycler==0.11.0
decorator==5.1.1
Deprecated==1.2.13
entrypoints==0.4
firebase-admin==5.4.0
fonttools==4.37.4
gcloud==0.18.3
gitdb==4.0.9
GitPython==3.1.27
google-api-core==2.10.1
google-api-python-client==2.64.0
google-auth==2.12.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.7.1
google-cloud-secret-manager==2.12.5
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
grpc-google-iam-v1==0.12.4
grpcio==1.49.1
grpcio-status==1.49.1
httplib2==0.20.4
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.9.0
Jinja2==3.1.2
jsonschema==4.16.0
jwcrypto==1.4.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.0
msgpack==1.0.4
numpy==1.23.3
oauth2client==4.1.3
packaging==21.3
pandas==1.5.0
Pillow==9.2.0
pkgutil-resolve-name==1.3.10
plotly==5.10.0
proto-plus==1.22.1
protobuf==3.20.3
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.15.0
pydeck==0.8.0b3
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
Pyrebase4==4.5.0
pyrsistent==0.18.1
python-dateutil==2.8.2
python-jwt==4.0.0
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
requests-toolbelt==0.9.1
rich==12.6.0
rsa==4.9
semver==2.13.0
six==1.16.0
smmap==5.0.0
streamlit==1.13.0
tenacity==8.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
typing-extensions==4.3.0
tzdata==2022.4
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.12
validators==0.20.0
watchdog==2.1.9
wrapt==1.14.1
zipp==3.8.1
